Bug fixes and results plotter - Added check for clearing the DB on start - Stream database for concepts instead of loading all into memory - Bug fix for checking start_id in id_to_uri and therefore not adding relations - Creating pickling file - Function for getting config - ResultsPlotter class for plotting clustering graph and calculating R² - Added averaged_perceptron_tagger_eng NLTK download

+++ build_ontology.py
@@ -16,6 +16,8 @@ import time
 from benchmarking.benchmark import Benchmark
 from clustering.cluster_concepts import ConceptClusterer
 from knowledge_bases import ConceptNetLoader
+from tools.config import get_config
+from tools.pickling import save_to_pickle, load_from_pickle
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 
@@ -31,6 +33,7 @@ class OntologyBuilder:
         self.config = config
         self.db_params = config['database']
         self.conn = None
+        self.ns = Namespace(self.config['turtle_export']['base_uri'])
 
         sources = self.config['general']['sources']
         mapping_types = ["edge", "pos"]
@@ -96,18 +99,21 @@ class OntologyBuilder:
         with self.conn.cursor() as cursor:
             try:
                 if self.config['general']['clear_db_on_start']:
-                    tables_to_delete = self.config['general']['tables_to_delete']
-                    sources_to_delete = self.config['general']['source_to_delete']
+                    user_input = input("Are you sure you want to clear the database? [y/n]")
 
-                    if len(self.config['general']['source_to_delete']) == 0:
-                        logging.info(f"Clearing existing tables ({', '.join(tables_to_delete)}) for all sources")
-                        for table in tables_to_delete:
-                            cursor.execute("DROP TABLE IF EXISTS %s CASCADE", [AsIs(table)])
-                    else:
-                        logging.info(f"Clearing existing tables ({', '.join(tables_to_delete)}) for sources: {self.config['general']['source_to_delete']}")
-                        for table in tables_to_delete:
-                            for source in sources_to_delete:
-                                cursor.execute("DELETE FROM %s WHERE source = %s", [AsIs(table), source])
+                    if user_input == 'y':
+                        tables_to_delete = self.config['general']['tables_to_delete']
+                        sources_to_delete = self.config['general']['source_to_delete']
+
+                        if len(self.config['general']['source_to_delete']) == 0:
+                            logging.info(f"Clearing existing tables ({', '.join(tables_to_delete)}) for all sources")
+                            for table in tables_to_delete:
+                                cursor.execute("DROP TABLE IF EXISTS %s CASCADE", [AsIs(table)])
+                        else:
+                            logging.info(f"Clearing existing tables ({', '.join(tables_to_delete)}) for sources: {self.config['general']['source_to_delete']}")
+                            for table in tables_to_delete:
+                                for source in sources_to_delete:
+                                    cursor.execute("DELETE FROM %s WHERE source = %s", [AsIs(table), source])
 
                 # Create tables
                 cursor.execute('''
@@ -162,8 +168,8 @@ class OntologyBuilder:
         WiktionaryLoader(self).load_data()
         logging.info("Ontology build process finished")
 
-    def _get_safe_uri(self, term, namespace):
-        return URIRef(namespace + quote(term)) if re.search(r'[^a-zA-Z0-9_-]', term) else namespace[term]
+    def _get_safe_uri(self, term):
+        return URIRef(self.ns + quote(term)) if re.search(r'[^a-zA-Z0-9_-]', term) else self.ns[term]
 
     def dump_to_turtle(self, file_path):
         start = time.time()
@@ -171,39 +177,51 @@ class OntologyBuilder:
         if not self.conn: logging.error("No DB connection for Turtle dump"); return
         logging.info(f"Dumping database to Turtle file: {file_path}")
 
-        g = Graph()
-        ns = Namespace(self.config['turtle_export']['base_uri'])
-        g.bind(self.config['turtle_export']['base_prefix'], ns)
-        g.bind("owl", OWL)
-        g.bind("rdfs", RDFS)
-        g.bind("xsd", XSD)
 
-        # Setup classes from JSON
-        self.create_classes(g, ns)
+        g = load_from_pickle('graph.pkl')
+        if g is None:
+            g = Graph()
+            g.bind(self.config['turtle_export']['base_prefix'], self.ns)
+            g.bind("owl", OWL)
+            g.bind("rdfs", RDFS)
+            g.bind("xsd", XSD)
 
-        # Add logical functions
-        ParmenidesLoader.add_logical_functions(self.config, g)
+            # Setup classes from JSON
+            self.create_classes(g)
+
+            # Add logical functions
+            ParmenidesLoader.add_logical_functions(self.config, g)
 
         all_properties = set()
-        with self.conn.cursor(name='concepts') as cursor:
-            cursor.execute("SELECT id, term, part_of_speech, source FROM concepts")
-            concepts_data = cursor.fetchall()
-            id_to_uri = {cid: self._get_safe_uri(term, ns) for cid, term, _, _ in concepts_data}
 
-            for cid, term, pos, source in tqdm(concepts_data, desc="Processing Concepts"):
-                if pos.lower() in self.rejected_classes: continue #or 'GeoNames' in source: continue
+        self.id_to_uri = load_from_pickle('id_to_uri.pkl')
+        if self.id_to_uri is None:
+            self.id_to_uri = {}
+            with self.conn.cursor() as count_cursor:
+                count_cursor.execute("SELECT COUNT(*) FROM concepts")
+                total_concepts = count_cursor.fetchone()[0]
 
-                uri = id_to_uri[cid]
+            with self.conn.cursor(name='concepts') as cursor:
+                cursor.execute("SELECT id, term, part_of_speech, source FROM concepts")
+                for cid, term, pos, source in tqdm(cursor, total=total_concepts, desc="Processing Concepts"):
+                    if pos.lower() in self.rejected_classes:
+                        continue
 
-                for i_pos in self.equivalent_classes.get(pos, [pos]):
-                    g.add((uri, RDF.type, ns[i_pos]))
-                g.add((uri, RDFS.label, Literal(term, datatype=XSD.string)))
+                    uri = self._get_safe_uri(term)
+                    self.id_to_uri[cid] = uri
+
+                    for i_pos in self.equivalent_classes.get(pos, [pos]):
+                        g.add((uri, RDF.type, self.ns[i_pos]))
+                    g.add((uri, RDFS.label, Literal(term, datatype=XSD.string)))
+
+            save_to_pickle('graph.pkl', g)
+            save_to_pickle('id_to_uri.pkl', self.id_to_uri)
 
         if not self.should_cluster:
             with self.conn.cursor(name='relations') as cursor:
                 cursor.execute("SELECT start_concept_id, end_concept_id, relation_type, weight, source FROM relations")
                 for start_id, end_id, rel_type, weight, source in tqdm(cursor, desc="Processing Relations"):
-                    self.add_relation_to_graph(g, start_id, end_id, id_to_uri, ns, rel_type, weight)
+                    self.add_relation_to_graph(g, start_id, end_id, rel_type, weight)
         else:
             with self.conn.cursor(name='cluster_relations') as cursor:
                 cursor.execute("SELECT start_cluster_id, end_cluster_id, relation_type, weight FROM cluster_relations")
@@ -212,21 +230,21 @@ class OntologyBuilder:
 
                     for start_id in start_ids:
                         for end_id in end_ids:
-                            self.add_relation_to_graph(g, start_id, end_id, id_to_uri, ns, rel_type, weight)
+                            self.add_relation_to_graph(g, start_id, end_id, rel_type, weight)
 
         with self.conn.cursor(name='properties') as cursor:
             cursor.execute("SELECT concept_id, type, value, source FROM properties")
             for concept_id, prop_type, value, source in tqdm(cursor, desc="Processing Properties"):
-                if concept_id in id_to_uri:
-                    concept_id = id_to_uri[concept_id]
-                    prop_uri = self._get_safe_uri(prop_type, ns)
+                if concept_id in self.id_to_uri:
+                    concept_id = self.id_to_uri[concept_id]
+                    prop_uri = self._get_safe_uri(prop_type)
                     g.add((concept_id, prop_uri, Literal(value)))
                     all_properties.add((prop_uri, OWL.DatatypeProperty))
 
         for prop_uri, prop_type in tqdm(all_properties, desc="Adding Properties"):
             g.add((prop_uri, RDF.type, prop_type))
 
-        self.add_pos_tag_classes(g, ns)
+        self.add_pos_tag_classes(g)
 
         try:
             logging.info("Serialising ontology")
@@ -247,12 +265,11 @@ class OntologyBuilder:
         except Exception as e:
             logging.error(f"Failed to write Turtle file: {e}")
 
-    def add_relation_to_graph(self, g, start_id, end_id, id_to_uri, ns, rel_type, weight):
-        if start_id in id_to_uri and end_id in id_to_uri:
-            start_uri, end_uri = id_to_uri[start_id], id_to_uri[end_id]
+    def add_relation_to_graph(self, g, start_id, end_id, rel_type, weight):
+        if start_id in self.id_to_uri.keys() and end_id in self.id_to_uri.keys():
+            start_uri, end_uri = self.id_to_uri[start_id], self.id_to_uri[end_id]
 
             mapping = self.full_mappings.get(rel_type.lower(), {'rel': rel_type, 'relNegated': False, 'swap': False})
-
             rel, is_negated, swap = mapping.get('rel'), mapping.get('relNegated', False), mapping.get('swap', False)
 
             sub_property_list = {'rel': rel, 'is_negated': Literal(is_negated), 'weight': Literal(weight)}
@@ -261,16 +278,16 @@ class OntologyBuilder:
             if sub_list_key in self.annotation_property_list:
                 rel_uri = self.annotation_property_list[sub_list_key]
             else:
-                rel_uri = self._get_safe_uri(f"{rel} {self.prop_id}", ns)
+                rel_uri = self._get_safe_uri(f"{rel} {self.prop_id}")
                 self.annotation_property_list[sub_list_key] = rel_uri
                 self.prop_id += 1
 
-            g.add((ns[rel], RDF.type, OWL.AnnotationProperty))
-            g.add((rel_uri, RDF.type, ns[rel]))
+            g.add((self.ns[rel], RDF.type, OWL.AnnotationProperty))
+            g.add((rel_uri, RDF.type, self.ns[rel]))
 
             for sub_prop_key, sub_prop_value in sub_property_list.items():
                 if sub_prop_key == 'rel': continue
-                g.add((rel_uri, ns[sub_prop_key], sub_prop_value))
+                g.add((rel_uri, self.ns[sub_prop_key], sub_prop_value))
 
             # p_uri = self._get_safe_uri(rel, ns)
             # all_properties.add((rel_uri, OWL.ObjectProperty))
@@ -278,45 +295,45 @@ class OntologyBuilder:
             s, t = (end_uri, start_uri) if swap else (start_uri, end_uri)
             g.add((s, rel_uri, t))
 
-    def create_classes(self, g: Graph, ns: Namespace):
+    def create_classes(self, g):
         logging.info("Creating ontology classes")
         for parent, children in self.ontology_classes.items():
             children = children['classes']
-            parent_uri = ns[parent]
+            parent_uri = self.ns[parent]
             g.add((parent_uri, RDF.type, OWL.Class))
-            self.create_sub_classes(parent, children, g, ns, parent_uri)
+            self.create_sub_classes(parent, children, g, parent_uri)
 
-    def create_sub_classes(self, parent, children, g: Graph, ns: Namespace, parent_uri: URIRef):
+    def create_sub_classes(self, parent, children, g: Graph, parent_uri: URIRef):
         ignored_keys = ['classes', 'sameAs']
         for child in children:
-            child_uri = ns[child]
+            child_uri = self.ns[child]
             g.add((child_uri, RDF.type, OWL.Class))
             g.add((child_uri, RDFS.subClassOf, parent_uri))
             if 'classes' in children[child]:
-                self.create_sub_classes(parent, children[child]['classes'], g, ns, child_uri)
+                self.create_sub_classes(parent, children[child]['classes'], g, child_uri)
             elif child not in ignored_keys and len(children[child]) > 0:
-                self.create_sub_classes(parent, children[child], g, ns, child_uri)
+                self.create_sub_classes(parent, children[child], g, child_uri)
 
             if len(children[child]) > 0 and 'sameAs' in children[child]:
                 self.equivalent_classes[child] = [child, children[child]['sameAs']]
 
-    def add_pos_tag_classes(self, g, ns):
+    def add_pos_tag_classes(self, g):
         logging.info("Adding POS tag classes")
-        g.add((ns['POSTag'], RDF.type, OWL.Class))
+        g.add((self.ns['POSTag'], RDF.type, OWL.Class))
 
         new_triples = set()
         for pos_tag, mapping in tqdm(self.pos_tag_mappings.items(), desc="Processing POS tagging"):
-            pos_uri = self._get_safe_uri(pos_tag, ns)
-            g.add((pos_uri, RDFS.subClassOf, ns['POSTag']))
+            pos_uri = self._get_safe_uri(pos_tag)
+            g.add((pos_uri, RDFS.subClassOf, self.ns['POSTag']))
 
             if "classes" not in mapping:
                 continue
 
             for class_name, class_details in mapping["classes"].items():
-                class_uri = ns[class_name]
+                class_uri = self.ns[class_name]
 
                 required_property_uris = [
-                    (self._get_safe_uri(prop, ns), Literal(True))
+                    (self._get_safe_uri(prop), Literal(True))
                     for prop in class_details.get("properties", [])
                 ]
 
@@ -344,20 +361,15 @@ class OntologyBuilder:
 
         with self.conn.cursor() as cursor:
             for concept_id in concept_ids:
-                cursor.execute("SELECT id, term, part_of_speech FROM concepts WHERE id = %s", [concept_id])
-                concepts.append(cursor.fetchone())
+                cursor.execute("SELECT id FROM concepts WHERE id = %s", [concept_id])
+                concepts.append(list(cursor.fetchone())[0])
         return concepts
 
     def close(self):
         if self.conn: self.conn.close(); logging.info("Database connection closed")
 
 def main(iterations = 1):
-    try:
-        with open('config.yaml', 'r') as f:
-            config = yaml.safe_load(f)
-    except FileNotFoundError:
-        logging.error("Configuration file 'config.yaml' not found")
-        exit(1)
+    config = get_config()
 
     builder = None
     try:
